+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 * Add logging of fetching changes
 * Fix new warnings from `rubocop` v0.89.0 update
 * Support of `markdownlint` in travis-ci
+* Increase test coverage to 100%
+* Add option to `GithubRepoChanges.new` to force read config from file
 
 ### Fixes